+++ src/kicad_tools/validate/rules/__init__.py
@@ -27,7 +27,7 @@ from .silkscreen import (
 from .single_pad_net import SinglePadNetRule
 from .solder_mask import SolderMaskPadRules
 from .via_in_pad import ViaInPadRule
-from .zone_fill import ZoneFillRule
+from .zone_fill import IsolatedCopperRule, ZoneFillRule
 
 __all__ = [
     "DRC_TOLERANCE",
@@ -43,6 +43,7 @@ __all__ = [
     "DimensionRules",
     "EdgeClearanceRule",
     "ImpedanceRule",
+    "IsolatedCopperRule",
     "MatchGroupLengthSkewRule",
     "NetImpedanceSpec",
     "SinglePadNetRule",